# feat(bb): optimize tuple creation