35% improvement in VM compilation time, and I'm sure it will improve BB's compilation time in general. Using this same trick, avoiding tuple concat and recursive usage, we can make other functions faster to compile as well.

BEFORE
```
VM Total Compilation time(ms): 684615
255329 ms: std::tuple_cat<$> (3092 times, avg 82 ms)
243861 ms: bb::create_sumcheck_tuple_of_tuples_of_univariates<$> (99 times, avg 2463 ms)
188579 ms: std::__tuple_concater<$>::_S_do<$> (3092 times, avg 60 ms)
118890 ms: bb::create_tuple_of_arrays_of_values<$> (177 times, avg 671 ms)
112903 ms: std::tuple<$> (6588 times, avg 17 ms)
```

AFTER
```
VM Total Compilation time(ms): 451748
41011 ms: bb::create_sumcheck_tuple_of_tuples_of_univariates<$> (99 times, avg 414 ms)
40975 ms: bb::create_sumcheck_tuple_of_tuples_of_univariates_internal<$> (99 times, avg 413 ms)
```